Rearrange folders and files in components
