+++ client/src/components/General/Login/index.jsx
@@ -1,0 +1,3 @@
+import Login from "./Component";
+
+export default Login;
